Split parser block helpers
